+++ src/Controller.php
@@ -31,18 +31,29 @@ class Controller
      *
      * If an error occurs, it is outputted
      *
-     * @param string $url       Router URL
-     * @param array  $resources List of resources available
-     * @param array  $config    Configurations for the Router
-     *                          @see Router::CONFIGURABLE
+     * @param string $url          Router URL
+     * @param array  $resources    List of resources available
+     * @param array  $config       Configurations for the Router
+     *                             @see Router::CONFIGURABLE
+     * @param string $router_class Router class to be used
+     *
+     * @throws \LogicException If $router_class is not subclass of Router
      */
     public function __construct(
         string $url,
         array $resources,
-        array $config = null
+        array $config = null,
+        string $router_class = null
     ) {
+        if ($router_class === null) {
+            $router_class = Router::class;
+        } elseif (!is_subclass_of($router_class, Router::class)) {
+            $message = $router_class . ' is not subclass of ' . Router::class;
+            throw new \LogicException($message);
+        }
+
         try {
-            $this->router = new Router($url, $resources, $config);
+            $this->router = new $router_class($url, $resources, $config);
         } catch (RouterException $e) {
             $e->getResponse()->output();
         }