+++ CHANGELOG.md
@@ -22,6 +22,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `Router::parseContentType()`
 - `Resouce->payload`
 - `Router::externalHandler()`
+- `Router::ENABLE_METHOD_OVERRIDE`
+- Extending section
 
 ### Changed
 - Update dependencies
@@ -35,6 +37,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Update `Router::parseBody()`
 - Update `Router::run()`
 - Route extensions are only used in `GET` and `HEAD` requests
+- Allow using a different Router in the Controller
+- `Router::requestRoot()` receive headers and request payload
 
 ### Deprecated